Делает ДЗ 5.1
